commiting model and view graphic communication
