+++ src/files/interfaces/Geometry.java
@@ -1,0 +1,7 @@
+package files.interfaces;
+
+import files.models.shapes.Point;
+
+public interface Geometry {
+    boolean isColliding(Point point);
+}
