+++ src/files/models/Shape.java
@@ -1,0 +1,6 @@
+package files.models;
+
+import files.interfaces.Geometry;
+
+public abstract class Shape implements Geometry {
+}
