+++ src/files/views/Graphic.java
@@ -1,9 +1,8 @@
 package files.views;
 
 import files.models.Graph;
-import files.views.Environment;
+import files.interfaces.Environment;
 import files.views.environment.Circle;
-import files.views.environment.Line;
 import files.views.environment.Text;
 import files.views.environment.Wall;
 import javafx.scene.canvas.GraphicsContext;
@@ -11,7 +10,6 @@ import javafx.scene.paint.Color;
 import javafx.scene.text.FontWeight;
 
 import java.util.ArrayList;
-import java.util.LinkedList;
 
 // TODO: remove this.draw() from Graphic constructor, implement Environment on Graphic class and use the draw method in View or UI
 public class Graphic {
