+++ src/files/views/environment/Circle.java
@@ -1,5 +1,5 @@
 package files.views.environment;
-import files.views.Environment;
+import files.interfaces.Environment;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Color;
